Compiler: Cygwin is identified by os(Cygwin) instead of os(Windows)
Since Mingw, Cygwin and MSVC are definitely vary greatly in ABI and C language level
behavior, it is neccessary to distinguish each other in Swift. (In target name,
MSVC is x86_64-*-windows-msvc, Cygwin is x86_64-*-windows-cygnus)

Before change, os(Windows) was *-*-windows-*, and now,
os(Cygwin) is for *-*-windows-cygnus, and os(Windows) is for *-*-windows-msvc.